DIALOGS :: Added custom function for displaying dialogs/alerts
The problem with Shopify form error and alert handling in Liquid is that 
with they're required to be enclosed in the Shopify form liquid tag (`{% 
form…`). This means that any content they render is within this form tag 
too. This becomes an issue when trying to display alerts elsewhere on 
the page because it must be absolutely position even if the client wants 
it to appear in the regular document flow.

The solution I've come up with for now is to capture errors and output 
them into a snippet that injects them into an existing dialog element 
using JavaScript near the top of the body. This <dialog> element is 
semantic and can be placed anywhere in the DOM.

The only issue is for some forms that redirect after submission, namely 
the reset password form. There is no way to capture the form message 
during these cases as well as method. 

There might be a way to do this with Shopify's Storefront API. I had 
trouble implementing it, unfortunatley (please see my workaround in the 
following commit). It's also worth asking whether we want to deal with 
an API and granting the developer its associated permissions for just 
one or two forms.